some changes in classes, blank UI

+++ MPD_McNaughton_algorithm/Task.cs
@@ -7,13 +7,13 @@ namespace MPD_McNaughton_algorithm
 {
     class Task
     {
-        public Task(String name, Int16 duration)
+        public Task(String name, Int32 duration)
         {
             Name = name;
             Duration = duration;
         }
 
         public String Name { get; private set; }
-        public Int16 Duration { get; private set; }
+        public Int32 Duration { get; private set; }
     }
 }
